% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_px_per_plate.R
\name{calculate_px_per_plate}
\alias{calculate_px_per_plate}
\title{Title Calculate the Price Per Plate by Recipe}
\usage{
calculate_px_per_plate(all_plates_df, recipes_df, prices)
}
\arguments{
\item{all_plates_df}{dataframe holding the total number of plates of recipe that will be made}

\item{recipes_df}{recipes dataframe downloaded from Ronalds googlesheet}

\item{prices}{historical prices of all the ingredients that go into Ronalds Recipes}
}
\value{

}
\description{
Title Calculate the Price Per Plate by Recipe
}
