% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shopping_list.R
\name{generate_shopping_list}
\alias{generate_shopping_list}
\title{Function to generate a shopping list for Ronald
The output is a dataframe}
\usage{
generate_shopping_list(all_plates_df, recipe_df)
}
\arguments{
\item{all_plates_df}{Dataframe holding the total number of plates of recipe that will be made}

\item{recipe_df}{Recipes dataframe downloaded from Ronalds googlesheet}
}
\value{

}
\description{
Function to generate a shopping list for Ronald
The output is a dataframe
}
