% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_data_from_sheet}
\alias{get_data_from_sheet}
\title{Get Recipes}
\usage{
get_data_from_sheet(recipe_ss_url, sheet_name)
}
\arguments{
\item{recipe_ss_url}{The url of the googlesheet - you must ensure that it is publicly readable}

\item{sheet_name}{The name of the sheet}
}
\value{

}
\description{
Get data from Ronalds Recipes Spreadsheet using googlesheets4
Ronalds Recipe Sheets hold all information relating to his catering business
}
